fix(parsers): surface parser failures instead of swallowing them
Review round 1 for B7. The new strict parsers were correct but their errors
were discarded at every production caller, so a format change still presented
as a plausible-looking empty or zeroed result.

- AppDelegate+Classification (single-sample and batch EsViritu): replace
  `(try? …) ?? []` with a do/catch that logs to OperationCenter at .error and
  rethrows. The single-sample path already fails the operation and shows an
  "EsViritu Failed" alert; the batch path already records the sample into
  failedResults. Both now report a detection-parse failure instead of an
  empty-but-successful run.
- MetagenomicsImportService: new `parseEsVirituVirusCount` propagates
  `MetagenomicsImportError.parseFailed` when the file HAS an EsViritu header but
  fails column validation, and keeps the raw `countDataRows` fallback only when
  no recognizable header is present. `EsVirituDetectionParser.isHeaderLine` is
  now public so importers can make that distinction.
- ProjectUniversalSearchIndex+Indexers: keeps best-effort `try?` semantics but
  logs at .error so an indexing-time format change is not invisible.
- FASTQDerivativeService (countFASTQReads, parseFASTQStats): still return nil,
  but log at .error first.
- FASTQDatasetViewController: a seqkit failure or parse failure now throws
  `FASTQQualityReportError.statisticsUnavailable` into the existing failure path
  (fails the operation, sets status, opens the Operations Panel) rather than
  rendering zeros as if they were measurements.
- SeqkitStatsRow.sumLen is now Int64; call-site widenings dropped.
- parseIdxstats is strict: a non-numeric length or count throws
  malformedIdxstatsRow instead of coercing to 0. populateFromIdxstats stays
  tolerant, as documented.

Tests: MetagenomicsImportService throws .parseFailed naming read_count for a
header missing that column, and still falls back to a row count for a headerless
partial export; SeqkitStatsParser handles a sum_len beyond the 32-bit range;
parseIdxstats throws on non-numeric count and length.

+++ Sources/LungfishApp/Views/DatabaseBrowser/DatabaseBrowserViewController.swift
@@ -3376,7 +3376,7 @@ public class DatabaseBrowserViewModel: ObservableObject {
 
         return SeqkitSummary(
             numSeqs: row.numSeqs,
-            sumLen: Int64(row.sumLen),
+            sumLen: row.sumLen,
             minLen: row.minLen,
             avgLen: row.avgLen,
             maxLen: row.maxLen,
